+++ app/jobs/sync_frilans_finans_user_job.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+class SyncFrilansFinansUserJob < ApplicationJob
+  def perform(user:)
+    SyncFrilansFinansUserService.call(user: user)
+  end
+end