+++ app/notifiers/accepted_applicant_confirmation_overdue_notifier.rb
@@ -5,8 +5,7 @@ class AcceptedApplicantConfirmationOverdueNotifier < BaseNotifier
 
     notify(locale: owner.locale) do
       JobMailer.
-        accepted_applicant_confirmation_overdue_email(job_user: job_user, owner: owner).
-        deliver_later
+        accepted_applicant_confirmation_overdue_email(job_user: job_user, owner: owner)
     end
   end
 end