+++ app/notifiers/invoice_failed_to_connect_to_frilans_finans_notifier.rb
@@ -7,8 +7,7 @@ class InvoiceFailedToConnectToFrilansFinansNotifier < BaseNotifier
       mailer_args = { user: user, ff_invoice: ff_invoice }
       notify do
         AdminMailer.
-          invoice_failed_to_connect_to_frilans_finans_email(**mailer_args).
-          deliver_later
+          invoice_failed_to_connect_to_frilans_finans_email(**mailer_args)
       end
     end
   end