+++ app/notifiers/job_cancelled_notifier.rb
@@ -6,8 +6,7 @@ class JobCancelledNotifier < BaseNotifier
 
       notify(locale: user.locale) do
         JobMailer.
-          job_cancelled_email(user: user, job: job).
-          deliver_later
+          job_cancelled_email(user: user, job: job)
       end
     end
   end