+++ app/notifiers/magic_login_link_notifier.rb
@@ -4,14 +4,12 @@ class MagicLoginLinkNotifier < BaseNotifier
   def self.call(user:)
     notify do
       UserTexter.
-        magic_login_link_text(user: user).
-        deliver_later
+        magic_login_link_text(user: user)
     end
 
     notify do
       UserMailer.
-        magic_login_link_email(user: user).
-        deliver_later
+        magic_login_link_email(user: user)
     end
   end
 end