+++ app/notifiers/new_job_comment_notifier.rb
@@ -7,8 +7,7 @@ class NewJobCommentNotifier < BaseNotifier
 
     notify(locale: owner.locale) do
       JobMailer.
-        new_job_comment_email(comment: comment, job: job).
-        deliver_later
+        new_job_comment_email(comment: comment, job: job)
     end
   end
 end