+++ app/notifiers/user_welcome_notifier.rb
@@ -3,8 +3,7 @@ class UserWelcomeNotifier < BaseNotifier
   def self.call(user:)
     notify do
       UserMailer.
-        welcome_email(user: user).
-        deliver_later
+        welcome_email(user: user)
     end
   end
 end